refactor: securityfilterChain bean moved to "SecurityConfiguration" class
