+++ server/src/main/java/com/app/server/configs/ApplicationConfiguration.java
@@ -7,15 +7,11 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
 
-import com.app.server.common.CONSTANT;
 import com.app.server.user.UserRepository;
 
 /**
@@ -79,29 +75,6 @@ public class ApplicationConfiguration {
   @Bean
   public AuthenticationManager authenticationManager(AuthenticationConfiguration config) throws Exception {
     return config.getAuthenticationManager();
-  }
-
-  @Bean
-  public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-    http
-        .csrf(customizer -> customizer.disable())
-        .sessionManagement(customizer -> customizer.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-        .authorizeHttpRequests(customizer -> {
-          customizer
-              .requestMatchers("/api/auth**")
-              .permitAll()
-
-              .requestMatchers("/api/admin**")
-              .hasRole(CONSTANT.ROLE_ADMINISTRATOR)
-
-              .requestMatchers("/api/users**")
-              .hasRole(CONSTANT.ROLE_USER)
-
-              .anyRequest()
-              .authenticated();
-        });
-
-    return http.build();
   }
 
 }
